Changed the data type of cashflow from float to decimal

+++ Fudge-It/Models/UserProfile.cs
@@ -13,7 +13,7 @@ namespace Fudge_It.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime CreatedDate { get; set; }
-        public float Cashflow { get; set; }
+        public decimal Cashflow { get; set; }
 
     }
 }
